requirements.txt
torch==2.7.0
numpy==2.2.4
spacy==3.8.5
sentence-transformers==4.1.0
faiss-cpu==1.11.0
flask==3.1.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
dotenv==0.9.9
# %%